requirements.txt
praw==7.7.1
python-dotenv==1.0.0
transformers==4.36.0
torch==2.2.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.2.0
schedule==1.2.1 